Feat : 내가 참여한 팟 리스트 조회 (Participating) API 구현 및 연동
