+++ presentation/src/main/java/com/pat/presentation/ui/map/MapScreen.kt
@@ -2,13 +2,13 @@ package com.pat.presentation.ui.map
 
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.runtime.Composable
-import com.pat.presentation.ui.proof.PattingScreenView
+import com.pat.presentation.ui.proof.ProofScreenView
 
 @Composable
 fun MapScreenView() {
     val scrollState = rememberScrollState()
 
 
-    PattingScreenView()
+    ProofScreenView()
 }
